Feature/admin

See merge request lbaw/lbaw2425/lbaw24153!26

+++ app/Http/Controllers/UserController.php
@@ -95,11 +95,17 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
+        $user = User::find($id);
+
+        if (!$user) {
+            return redirect()->route('admin.dashboard')->with('error', 'User not found.');
+        }
+
         $user->delete();
 
-        return redirect()->route('users.index')->with('success', 'User deleted successfully!');
+        return redirect()->route('admin.dashboard')->with('success', 'User deleted successfully.');
     }
+
 
     public function score($id)
     {
